Add BeagleBone 720MHz support
Modified evm.c to provide basic support for TPS65217 PMIC that is
used on BeagleBone.  The code increases the USB current limit to
1300mA, sets the DCDC2 voltage to 1.275V, and the MPU frequency
to 720MHz.

If the TPS65217 does not respond to i2c_probe(), we assume that we
are on the EVM.  This patch was tested to ensure that 720MHz
support continues to work on the EVM as well.

Signed-off-by: Greg Guyotte <gguyotte@ti.com>