omap3: Do not initialize DDR memory from u-boot
Currently u-boot tries to initialize the second bank of DDR on
SDRC CS1, if attached. But this piece of code also hard codes
the size of DDR to 128MiB, thus limiting the total memory available
to the system.
This patch removes additional DDR memory initialization from u-boot
and stick with the settings done from the boot-strap.

Signed-off-by: Arun Joseph <arunjoseph@ti.com>