omap-common: Remove omap_boot_mode check from spl_nand_load_image
We do not yet have a ROM that passes useful information to us about how
the image was read from NAND so remove a check for doing that until we
have hardware that can tell us such.  At that point I suspect the check
will need to be handled as part of the board callback made from
nand_init.

Signed-off-by: Tom Rini <trini@ti.com>