Fastboot: Move boot device selection to board file
This u-boot has support for Fastboot on NAND flash or Fastboot on eMMC,
but not both.

The CONFIG_STORAGE_EMMC and FASTBOOT_PORT_OMAPZOOM_NAND_FLASHING
macros are defined in the same board file (only for AM335x),
in order for the selection of one macro between  the two modes.

By default FASTBOOT_PORT_OMAPZOOM_NAND_FLASHING is enabled, so fastboot
will use NAND device.

Change-Id: I3d228bab05ab5285f8e3678073f9d1da112fcbe2
Signed-off-by: Umakanta Patro <umakanta.patro@ti.com>